requirements.txt
blis==0.7.11
bokeh==3.5.2
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
cloudpickle==3.0.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.0
cymem==2.0.8
dask==2024.8.2
dask-expr==1.1.13
distributed==2024.8.2
dnspython==2.6.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl#sha256=6a0f857a2b4d219c6fa17d455f82430b365bf53171a2d919b9376e5dc9be032e
fsspec==2024.6.1
idna==3.8
importlib_metadata==8.4.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
locket==1.0.0
lz4==4.3.3
marisa-trie==1.2.0
MarkupSafe==2.1.5
msgpack==1.0.8
murmurhash==1.0.10
numpy==1.26.4
packaging==24.1
pandas==2.2.2
partd==1.4.2
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.4.0
preshed==3.0.9
psutil==6.0.0
pyarrow==17.0.0
pydantic==1.10.18
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
smart-open==6.4.0
sortedcontainers==2.4.0
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tblib==3.0.0
thinc==8.2.5
threadpoolctl==3.5.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
typer==0.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
xyzservices==2024.9.0
zict==3.0.0
zipp==3.20.1
